requirements.txt
django==3.0.7
boostrap4
django_filter==2.3.0
django_widget_tweaks==1.4.8
